+++ expo/ch1_basic/assets/adaptive-icon.png
