+++ expo/ch1_basic/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ch1-basics",
+  "name": "basic",
   "version": "1.0.0",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
